+++ pkg/asset/reader.go
@@ -6,6 +6,7 @@ type ScenarioReader interface {
 	Asset(templatePath string) ([]byte, error)
 	// List all available assets in the data source
 	// with a prefix of one of the name in the files array
+	// if the prefix array is nil, the prefix will not be tested
 	// excluding the name in the excluded array
 	AssetNames(prefixes, excluded []string, headerFile string) ([]string, error)
 }